Update dockerfile to more closely emulate real environment
As per the Buildpack API (https://devcenter.heroku.com/articles/buildpack-api)
the buildpack scripts are run in a environment stripped of most
environment variables - emulate that here by using `env -i`.

In real life there is a whitelist of env vars. Rather than put them all
in, by trial-and-error I determined only PATH and HOME are required.
It feels a bit hacky to do it like that but I think it's ok for now.