Bring in-dyno redis in line with postgres (#18)
There are now some languages that are relying on redis to be up and the `REDIS_URL` to be available during build time. The [postgres in-dyno db](https://github.com/heroku/heroku-buildpack-ci-postgresql/blob/master/bin/compile) as a few things to handle this that I've brought to the redis in-dyno buildpack. Mainly the `background` file is used to start the redis-server before each buildpack's compile is executed. I also went ahead and wrote the `REDIS_URL` to `export` to align with the postgres version.