Update default version aliases to latest minor/patch versions
Users of the buildpack can either:
- not specify a Redis version at all (in which case `6` is used for new apps,
  or the previously used version for existing apps)
- specify a major Redis version (such as `6` or `4`)
- specify an exact Redis version (such as `6.0.16`)

The current major version aliases are slightly outdated, so this PR updates
them to the latest minor/patch versions to ensure CI is using the same
versions as provisioned by the Heroku Redis addon.

In addition, the hope is that the newer 5.x version will fix the compile errors
seen under Heroku-22 in #27.

The alias version changes are as follows:
- `3`: `3.2.12` -> `3.2.13`
- `5`: `5.0.6` -> `5.0.14`
- `6`: `6.0.9` -> `6.2.7`

This also means the default CI Redis version for new apps that don't specify
a Redis version is now 6.2.x instead of 6.0.x - which matches the default for
Heroku Redis:
https://devcenter.heroku.com/articles/heroku-redis#version-support-and-legacy-infrastructure

GUS-W-10346688.