feat: overhaul - add tls, version support
This is a pretty huge monster of a change and needs to be split up.

In summary, to align with the Redis Version Support Policy from Heroku,
we need to drop all versions < 6.2.

Additionally, we need to support TLS to align with platform standards.
This is enabled by passing in `REDIS_CI_TLS` into the environment, which
will start a TLS-enabled `redis-server` instead of a plaintext one. This
allows for closer parity with production in CI.